requirements.txt
tensorflow-macos
tensorflow-metal
matplotlib
numpy 
pandas
